feat(habit-tracker-repository): add habit-event list method #18
