+++ app/src/main/java/pl/gawor/android/tayckner/repository/token.kt
@@ -1,0 +1,4 @@
+package pl.gawor.android.tayckner.repository
+
+const val JWT_TOKEN = "Bearer " +
+        "eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2NDAxODE4NzgsImV4cCI6MTY0MDE5MTQwMCwidXNlcklkIjozLCJ1c2VybmFtZSI6IndhemEifQ.r0XEKG_0KKP7drCaH8QuSs-V8A2-tCKM_VWqZL7oFns"
